with_params => impl & fixes
